renamed pump feature application sub-folder
